requirements.txt
--extra-index-url https://download.pytorch.org/whl/lts/1.8/cpu
antspyx==0.4.2 --only-binary=antspyx
antspynet==0.2.3
gast<=0.4.0,>=0.2.1
matplotlib==3.5.1
mo-dots==9.147.22086
nibabel==3.2.2
numpy==1.21.6
psutil==5.9.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
setproctitle==1.2.3
tensorflow_probability==0.19.0
torchsummary==1.5.1
torch==1.8.2
torchvision==0.9.2